pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-dataloader"
version = "1.0.0"
description = "Universal DataLoader - AI-Powered Knowledge Graph Construction"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "data-loading",
    "knowledge-graph", 
    "neo4j",
    "langchain",
    "ai",
    "nlp",
    "data-extraction",
    "graph-database",
    "llm",
    "rag"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Data Engineers",
    "Intended Audience :: System Administrators", 
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.9",
    "langchain-community>=0.3.9", 
    "langchain-openai>=0.3.9",
    "langchain-experimental>=0.3.3",
    "langchain-neo4j>=0.4.0",
    "neo4j>=5.25.0",
    "openai>=1.13.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.5.0",
    "typing-extensions>=4.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.991"
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0"
]
full = [
    "unstructured>=0.10.0",
    "pygrok>=1.0.0", 
    "pypdf>=3.0.0",
    "beautifulsoup4>=4.11.0",
    "markdown>=3.4.0",
    "langchain-text-splitters>=0.3.0"
]

[project.urls]
Homepage = "https://github.com/rrbanda/dataloader"
Repository = "https://github.com/rrbanda/dataloader"
Documentation = "https://github.com/rrbanda/dataloader/blob/main/README.md"
"Bug Reports" = "https://github.com/rrbanda/dataloader/issues"

[project.scripts]
universal-dataloader = "core.unified_dataloader:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "config*", "utils*", "examples*", "tests*"]

[tool.setuptools.package-data]
config = ["*.yaml", "*.yml"]
examples = ["*.py"]  
tests = ["*.py"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
